Update swagger spec
Universal use of value and values to describe options on a dimension has
caused a lot of confusion, so a lot of the work has been to replace the
field value with option or options.

Secondly the data model for request body and response body are now split
out into separate models.