Rework api to use a filter blueprint resource
Filter blueprint resource can be updated and submitted as many times as
the user wants. On submission a filter output resource is created and
will be updated with download urls by the backend export process. When
submitting a filter blueprint, the response body will return a link to
the filter output resource.

Also include unit test coverage.