Fix GET /filter-outputs/{id} endpoint
Filtering on CMD datasets is currently broken in develop. Once selections
are made by the user and the button is pressed to generate a download
an error is returned.

An error screen is displayed. When looking for the UUID in the URL bar
at the top of the screen in Kibana the below error is returned.

```
"{ "status_code": 500, "response": "failed to get dataset: filter blueprint not found", "error": "failed to get dataset: filter blueprint not found" }"
```

This is due to the `assert.FilterType(...)` being used instead of the
`assert.FilterOutputHandler(...)`

Resolves: 5669